refactor(assertions): isolate metric name rendering
