+++ src/assertions/index.ts
@@ -79,6 +79,7 @@ import { handlePython } from './python';
 import { handleRedteam } from './redteam';
 import { handleIsRefusal } from './refusal';
 import { handleRegex } from './regex';
+import { renderMetricName } from './renderMetricName';
 import { handleRougeScore } from './rouge';
 import { handleRuby } from './ruby';
 import { handleSearchRubric } from './searchRubric';
@@ -102,6 +103,8 @@ import { handleWebhook } from './webhook';
 import { handleWordCount } from './wordCount';
 import { handleIsXml } from './xml';
 
+export { renderMetricName };
+
 import type {
   AssertionOrSet,
   AssertionParams,
@@ -118,6 +121,7 @@ const DEFAULT_TRACE_FETCH_STABLE_POLLS = 2;
 const MAX_TRACE_FETCH_MAX_ATTEMPTS = 30;
 const MAX_TRACE_FETCH_RETRY_DELAY_MS = 5000;
 const MAX_TRACE_FETCH_STABLE_POLLS = 10;
+const nunjucks = getNunjucksEngine();
 
 export const MODEL_GRADED_ASSERTION_TYPES = new Set<AssertionType>([
   'agent-rubric',
@@ -314,35 +318,6 @@ const ASSERTION_HANDLERS: Record<
   webhook: handleWebhook,
   'word-count': handleWordCount,
 };
-
-const nunjucks = getNunjucksEngine();
-
-/**
- * Renders a metric name template with test variables.
- * @param metric - The metric name, possibly containing Nunjucks template syntax
- * @param vars - The test variables to use for rendering
- * @returns The rendered metric name, or the original if rendering fails
- */
-export function renderMetricName(
-  metric: string | undefined,
-  vars: Record<string, unknown>,
-): string | undefined {
-  if (!metric) {
-    return metric;
-  }
-  try {
-    const rendered = nunjucks.renderString(metric, vars);
-    if (rendered === '' && metric !== '') {
-      logger.debug(`Metric template "${metric}" rendered to empty string`);
-    }
-    return rendered;
-  } catch (error) {
-    logger.warn(
-      `Failed to render metric template "${metric}": ${error instanceof Error ? error.message : error}`,
-    );
-    return metric;
-  }
-}
 
 /**
  * Tests whether an assertion is inverse e.g. "not-equals" is inverse of "equals"
