+++ src/util/namedMetrics.ts
@@ -1,4 +1,4 @@
-import { renderMetricName } from '../assertions/index';
+import { renderMetricName } from '../assertions/renderMetricName';
 
 import type { GradingResult, Vars } from '../types/index';
